requirements.txt
Flask==2.3.2
requests==2.31.0
cryptography==41.0.3
pycryptodome==3.18.0
websockets==11.0.3
pyotp==2.8.0
py_ecc==6.1.0
mailjet-rest==1.3.4
phe==1.5.0
schedule
pip install flask-session
pip install paililer